backend/requirements.txt
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.3.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.4.2
langdetect==1.0.9
lxml==5.3.2
lxml_html_clean==0.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.0.2
packaging==24.2
pillow==11.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
safetensors==0.5.3
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.6
sympy==1.13.1
tinysegmenter==0.3
tldextract==5.2.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.1
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
